Initial commit of mysql-workbench. Fixes T243.
Summary:
Depends on D3052.

NOTES: kudos for this patch is due to @herbmillerjr. As there has been no response from him since november 2017, and expecting that many users can be interested in having mysql-workbench in the repos, I'm submitting this with the changes demanded by @JoshStrobl in D938. Summary of changes made:
- Built using the latest sources atm: 6.3.10 version.
- Using %apply_patches% for multiple patching.
- Remove sed and use patching instead.
- Removed unnecessary patch with the new sources: 007-proj.patch.

Test Plan:
- Create and save new queries.
- Create a table and insert 5000 rows. Select all the new inserted rows.
- Create and save EER Diagrams.

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake, hubris, JoshStrobl, herbmillerjr

Differential Revision: https://dev.solus-project.com/D3053